requirements.txt
streamlit
python-dotenv
pdfplumber
openai
numpy
faiss-cpu
scikit-learn
streamlit-extras
langchain
reportlab
